+++ system-tests/end2end-tests/src/test/java/org/eclipse/edc/catalog/EndToEndCatalogTest.java
@@ -9,7 +9,6 @@ import org.eclipse.edc.connector.contract.spi.types.offer.ContractOffer;
 import org.eclipse.edc.junit.extensions.EdcExtension;
 import org.eclipse.edc.spi.message.MessageContext;
 import org.eclipse.edc.spi.message.RemoteMessageDispatcher;
-import org.eclipse.edc.util.TestFunctions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -30,14 +29,14 @@ import static java.util.concurrent.CompletableFuture.completedFuture;
 import static java.util.stream.IntStream.range;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
-import static org.eclipse.edc.matchers.CatalogRequestMatcher.sentTo;
-import static org.eclipse.edc.util.TestFunctions.catalogBuilder;
-import static org.eclipse.edc.util.TestFunctions.catalogOf;
-import static org.eclipse.edc.util.TestFunctions.createOffer;
-import static org.eclipse.edc.util.TestFunctions.emptyCatalog;
-import static org.eclipse.edc.util.TestFunctions.insertSingle;
-import static org.eclipse.edc.util.TestFunctions.queryCatalogApi;
-import static org.eclipse.edc.util.TestFunctions.randomCatalog;
+import static org.eclipse.edc.catalog.TestFunctions.catalogBuilder;
+import static org.eclipse.edc.catalog.TestFunctions.catalogOf;
+import static org.eclipse.edc.catalog.TestFunctions.createOffer;
+import static org.eclipse.edc.catalog.TestFunctions.emptyCatalog;
+import static org.eclipse.edc.catalog.TestFunctions.insertSingle;
+import static org.eclipse.edc.catalog.TestFunctions.queryCatalogApi;
+import static org.eclipse.edc.catalog.TestFunctions.randomCatalog;
+import static org.eclipse.edc.catalog.matchers.CatalogRequestMatcher.sentTo;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.ArgumentMatchers.eq;
