+++ system-tests/end2end-tests/src/test/java/org/eclipse/edc/catalog/matchers/CatalogRequestMatcher.java
@@ -1,4 +1,4 @@
-package org.eclipse.edc.matchers;
+package org.eclipse.edc.catalog.matchers;
 
 import org.eclipse.edc.catalog.spi.CatalogRequest;
 import org.mockito.ArgumentMatcher;
