+++ system-tests/end2end-tests/src/test/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-org.eclipse.edc.instrumentation.MockInjectionExtension
+org.eclipse.edc.catalog.instrumentation.MockInjectionExtension
